requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.10
branca==0.6.0
breadability==0.1.20
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
confection==0.1.1
contourpy==1.1.0
cssselect==1.2.0
cssutils==2.7.1
cvzone==1.6.1
cycler==0.11.0
cymem==2.0.7
decorator==4.4.2
deprecation==2.1.0
docopt==0.6.2
docx2pdf==0.1.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
Flask==2.3.3
folium==0.14.0
fonttools==4.42.0
frozenlist==1.4.0
gTTS==2.4.0
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.9
img2pdf==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
moviepy==1.0.3
multidict==6.0.4
murmurhash==1.0.9
mysql==0.0.3
mysqlclient==2.2.0
networkx==3.1
nltk==3.8.1
numpy==1.25.2
openai==0.27.9
opencage==2.3.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.1
pathy==0.10.2
phonenumbers==8.13.21
pikepdf==8.6.0
Pillow==10.1.0
premailer==3.10.0
preshed==3.0.8
proglog==0.1.10
pycountry==22.3.5
pydantic==2.2.1
pydantic_core==2.6.1
pygame==2.5.1
PyMySQL==1.1.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqt5-tools==5.15.9.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
qrcode==7.4.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
regex==2023.8.8
requests==2.31.0
scipy==1.11.1
seaborn==0.13.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.0
srsly==2.4.7
sumy==0.11.0
thinc==8.1.12
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wasabi==1.1.2
Werkzeug==2.3.7
xlrd==2.0.1
yagmail==0.15.293
yarl==1.9.2
